requirements.txt
openai
supermemory
python-dotenv
